Revert "feat(firmware): add Rust product test bench (#10605)"
This reverts commit fb3718d10047bfd24b26f2fc7d9fc26671aa2a04.
