--- omi/firmware/bsim/Kconfig
@@ -1,1 +1,0 @@
-rsource "../omi/Kconfig"
