--- omi/firmware/bsim/client/CMakeLists.txt
@@ -1,4 +1,0 @@
-cmake_minimum_required(VERSION 3.20.0)
-find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
-project(omi_bsim_client)
-target_sources(app PRIVATE src/main.c)
