--- omi/firmware/bsim/client/prj.conf
@@ -1,8 +1,0 @@
-CONFIG_BT=y
-CONFIG_BT_CENTRAL=y
-CONFIG_BT_GATT_CLIENT=y
-CONFIG_BT_MAX_CONN=1
-CONFIG_LOG=y
-CONFIG_LOG_DEFAULT_LEVEL=3
-CONFIG_PRINTK=y
-CONFIG_MAIN_STACK_SIZE=4096
