--- omi/firmware/bsim/prj.conf
@@ -1,35 +1,0 @@
-CONFIG_BT=y
-CONFIG_BT_PERIPHERAL=y
-CONFIG_BT_DEVICE_NAME="Omi"
-CONFIG_BT_MAX_CONN=1
-CONFIG_BT_GATT_DYNAMIC_DB=y
-CONFIG_BT_GATT_CLIENT=y
-CONFIG_BT_L2CAP_TX_MTU=247
-CONFIG_BT_CONN_TX_MAX=20
-CONFIG_BT_L2CAP_TX_BUF_COUNT=20
-CONFIG_BT_BUF_ACL_RX_SIZE=251
-CONFIG_BT_BUF_ACL_TX_SIZE=251
-CONFIG_BT_BUF_ACL_TX_COUNT=10
-CONFIG_BT_ATT_TX_COUNT=20
-CONFIG_BT_USER_DATA_LEN_UPDATE=y
-CONFIG_BT_USER_PHY_UPDATE=y
-CONFIG_BT_AUTO_DATA_LEN_UPDATE=y
-CONFIG_BT_AUTO_PHY_UPDATE=y
-CONFIG_BT_GATT_AUTO_UPDATE_MTU=y
-CONFIG_BT_SETTINGS=y
-CONFIG_SETTINGS=y
-CONFIG_SETTINGS_NVS=y
-CONFIG_NVS=y
-CONFIG_FLASH=y
-CONFIG_FLASH_MAP=y
-CONFIG_RING_BUFFER=y
-CONFIG_LOG=y
-CONFIG_LOG_DEFAULT_LEVEL=3
-CONFIG_PRINTK=y
-CONFIG_MAIN_STACK_SIZE=4096
-CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=4096
-CONFIG_HEAP_MEM_POOL_SIZE=40000
-CONFIG_OMI_CODEC_OPUS=y
-CONFIG_OMI_ENABLE_BUTTON=y
-CONFIG_OMI_ENABLE_BUTTON_INPUT=n
-CONFIG_OMI_ENABLE_RFSW_CTRL=n
