--- omi/firmware/bsim/sysbuild.conf
@@ -1,1 +1,0 @@
-SB_CONFIG_PARTITION_MANAGER=n
