--- omi/firmware/bsim/sysbuild/hci_ipc.overlay
@@ -1,13 +1,0 @@
-&bt_hci_controller {
-	status = "okay";
-};
-
-&bt_hci_sdc {
-	status = "disabled";
-};
-
-/ {
-	chosen {
-		zephyr,bt-hci = &bt_hci_controller;
-	};
-};
