+++ omi/firmware/omi/src/lib/core/button.c
@@ -1,5 +1,10 @@
 #include "button.h"
 
+#include <zephyr/bluetooth/bluetooth.h>
+#include <zephyr/bluetooth/gatt.h>
+#include <zephyr/bluetooth/l2cap.h>
+#include <zephyr/bluetooth/services/bas.h>
+#include <zephyr/bluetooth/uuid.h>
 #include <zephyr/drivers/gpio.h>
 #include <zephyr/input/input.h>
 #include <zephyr/kernel.h>
@@ -8,12 +13,13 @@
 #include <zephyr/sys/poweroff.h>
 
 #include "haptic.h"
-#include "imu.h"
 #include "led.h"
 #include "mic.h"
 #include "speaker.h"
 #include "transport.h"
 #include "wdog_facade.h"
+
+#include "imu.h"
 #ifdef CONFIG_OMI_ENABLE_OFFLINE_STORAGE
 #include "sd_card.h"
 #endif
@@ -22,6 +28,41 @@ LOG_MODULE_REGISTER(button, CONFIG_LOG_DEFAULT_LEVEL);
 
 extern bool is_off;
 
+static void button_ccc_config_changed_handler(const struct bt_gatt_attr *attr, uint16_t value);
+static ssize_t button_data_read_characteristic(struct bt_conn *conn,
+                                               const struct bt_gatt_attr *attr,
+                                               void *buf,
+                                               uint16_t len,
+                                               uint16_t offset);
+
+static struct bt_uuid_128 button_uuid =
+    BT_UUID_INIT_128(BT_UUID_128_ENCODE(0x23BA7924, 0x0000, 0x1000, 0x7450, 0x346EAC492E92));
+static struct bt_uuid_128 button_characteristic_data_uuid =
+    BT_UUID_INIT_128(BT_UUID_128_ENCODE(0x23BA7925, 0x0000, 0x1000, 0x7450, 0x346EAC492E92));
+
+static struct bt_gatt_attr button_service_attr[] = {
+    BT_GATT_PRIMARY_SERVICE(&button_uuid),
+    BT_GATT_CHARACTERISTIC(&button_characteristic_data_uuid.uuid,
+                           BT_GATT_CHRC_READ | BT_GATT_CHRC_NOTIFY,
+                           BT_GATT_PERM_READ,
+                           button_data_read_characteristic,
+                           NULL,
+                           NULL),
+    BT_GATT_CCC(button_ccc_config_changed_handler, BT_GATT_PERM_READ | BT_GATT_PERM_WRITE),
+};
+
+static struct bt_gatt_service button_service = BT_GATT_SERVICE(button_service_attr);
+
+static void button_ccc_config_changed_handler(const struct bt_gatt_attr *attr, uint16_t value)
+{
+    if (value == BT_GATT_CCC_NOTIFY) {
+        LOG_INF("Client subscribed for notifications");
+    } else if (value == 0) {
+        LOG_INF("Client unsubscribed from notifications");
+    } else {
+        LOG_ERR("Invalid CCC value: %u", value);
+    }
+}
 static const struct device *const buttons = DEVICE_DT_GET(DT_ALIAS(buttons));
 static const struct gpio_dt_spec usr_btn = GPIO_DT_SPEC_GET_OR(DT_NODELABEL(usr_btn), gpios, {0});
 
@@ -47,6 +88,7 @@ static FSM_STATE_T current_button_state = IDLE;
 static uint32_t inc_count_1 = 0;
 static uint32_t inc_count_0 = 0;
 
+static int final_button_state[2] = {0, 0};
 const static int threshold = 10;
 
 static void reset_count()
@@ -56,34 +98,52 @@ static void reset_count()
 }
 static inline void notify_press()
 {
+    final_button_state[0] = BUTTON_PRESS;
     LOG_INF("Button pressed");
-    button_notify(BUTTON_PRESS);
+    struct bt_conn *conn = get_current_connection();
+    if (conn != NULL) {
+        bt_gatt_notify(conn, &button_service.attrs[1], &final_button_state, sizeof(final_button_state));
+    }
 }
 
 static inline void notify_unpress()
 {
+    final_button_state[0] = BUTTON_RELEASE;
     LOG_INF("Button released");
-    button_notify(BUTTON_RELEASE);
+    struct bt_conn *conn = get_current_connection();
+    if (conn != NULL) {
+        bt_gatt_notify(conn, &button_service.attrs[1], &final_button_state, sizeof(final_button_state));
+    }
 }
 
 static inline void notify_tap()
 {
+    final_button_state[0] = SINGLE_TAP;
     LOG_INF("Button single tap");
-    button_notify(SINGLE_TAP);
+    struct bt_conn *conn = get_current_connection();
+    if (conn != NULL) {
+        bt_gatt_notify(conn, &button_service.attrs[1], &final_button_state, sizeof(final_button_state));
+    }
 }
 
 static inline void notify_double_tap()
 {
-    // button press
+    final_button_state[0] = DOUBLE_TAP; // button press
     LOG_INF("Button double tap");
-    button_notify(DOUBLE_TAP);
+    struct bt_conn *conn = get_current_connection();
+    if (conn != NULL) {
+        bt_gatt_notify(conn, &button_service.attrs[1], &final_button_state, sizeof(final_button_state));
+    }
 }
 
 static inline void notify_long_tap()
 {
-    // button press
+    final_button_state[0] = LONG_TAP; // button press
     LOG_INF("Button long tap");
-    button_notify(LONG_TAP);
+    struct bt_conn *conn = get_current_connection();
+    if (conn != NULL) {
+        bt_gatt_notify(conn, &button_service.attrs[1], &final_button_state, sizeof(final_button_state));
+    }
 }
 
 #define BUTTON_PRESSED 1
@@ -197,6 +257,17 @@ void check_button_level(struct k_work *work_item)
     return 0;
 }
 
+static ssize_t button_data_read_characteristic(struct bt_conn *conn,
+                                               const struct bt_gatt_attr *attr,
+                                               void *buf,
+                                               uint16_t len,
+                                               uint16_t offset)
+{
+    LOG_INF("button_data_read_characteristic");
+    LOG_PRINTK("was_pressed: %d\n", final_button_state[0]);
+    return bt_gatt_attr_read(conn, attr, buf, len, offset, &final_button_state, sizeof(final_button_state));
+}
+
 static struct gpio_callback button_cb_data;
 
 static void button_gpio_callback(const struct device *dev, struct gpio_callback *cb, uint32_t pins)
@@ -262,6 +333,11 @@ int button_init()
 void activate_button_work()
 {
     k_work_schedule(&button_work, K_MSEC(BUTTON_CHECK_INTERVAL));
+}
+
+void register_button_service()
+{
+    bt_gatt_service_register(&button_service);
 }
 
 FSM_STATE_T get_current_button_state()
@@ -344,6 +420,7 @@ void turnoff_all()
         return;
     }
 
+    
     /* Persist an IMU timestamp base so we can estimate time across system_off. */
     lsm6dsl_time_prepare_for_system_off();
     k_msleep(1000);
