--- omi/firmware/scripts/devkit/OmiSimulator/.gitignore
@@ -1,1 +1,0 @@
-target/
