--- omi/firmware/scripts/devkit/OmiSimulator/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "omi-product-emulator"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-publish = false
-
-[dependencies]
-btleplug = "0.11"
-ctrlc = "3"
-futures = "0.3"
-omi-firmware-core = { path = "firmware-core" }
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-tokio = { version = "1", features = ["rt-multi-thread", "time"] }
-zip = { version = "4", default-features = false, features = ["deflate"] }
-
-[target.'cfg(target_os = "macos")'.dependencies]
-crepuscularity-gpui = { version = "0.5.8", features = ["macos"] }
-gpui = { version = "=0.2.2", default-features = false, features = ["font-kit", "macos-blade"] }
-
-[target.'cfg(target_os = "linux")'.dependencies]
-crepuscularity-gpui = { version = "0.5.8", features = ["x11"] }
-dbus = { version = "0.9", features = ["vendored"] }
-gpui = { version = "=0.2.2", default-features = false, features = ["font-kit", "x11"] }
-
-[target.'cfg(target_os = "windows")'.dependencies]
-crepuscularity-gpui = { version = "0.5.8", features = ["windows"] }
-gpui = { version = "=0.2.2", default-features = false, features = ["font-kit", "windows-manifest"] }
-
-[dev-dependencies]
-tempfile = "3"
