--- omi/firmware/scripts/devkit/OmiSimulator/firmware-core/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "omi-firmware-core"
-version = "0.1.0"
-edition = "2021"
-license = "MIT"
-publish = false
-
-[dependencies]
-serde = { version = "1", default-features = false, features = ["derive"] }
-
-[lib]
-path = "src/lib.rs"
