--- omi/firmware/scripts/devkit/OmiSimulator/native/macos/Info.plist
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>NSBluetoothAlwaysUsageDescription</key>
-    <string>Advertise an Omi test peripheral for local firmware and app validation.</string>
-</dict>
-</plist>
