--- omi/firmware/scripts/devkit/OmiSimulator/qemu/.cargo/config.toml
@@ -1,6 +1,0 @@
-[build]
-target = "thumbv7em-none-eabihf"
-
-[target.thumbv7em-none-eabihf]
-runner = "qemu-system-arm -machine mps2-an386 -cpu cortex-m4 -nographic -semihosting-config enable=on,target=native -kernel"
-rustflags = ["-C", "link-arg=-Tlink.x"]
