--- omi/firmware/scripts/devkit/OmiSimulator/qemu/build.rs
@@ -1,6 +1,0 @@
-fn main() {
-    println!(
-        "cargo:rustc-link-search={}",
-        std::env::var("CARGO_MANIFEST_DIR").unwrap()
-    );
-}
