--- omi/firmware/scripts/devkit/OmiSimulator/qemu/memory.x
@@ -1,5 +1,0 @@
-MEMORY
-{
-  FLASH : ORIGIN = 0x00000000, LENGTH = 4M
-  RAM : ORIGIN = 0x20000000, LENGTH = 4M
-}
