--- omi/firmware/scripts/devkit/OmiSimulator/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod core;
